Move FrontJsonFapi.scala from facia to common
We want to be able to also use it in `article/app/agents/CuratedContentAgent.scala`
