+++ facia/app/controllers/FaciaControllers.scala
@@ -1,10 +1,10 @@
 package controllers
 
 import com.softwaremill.macwire._
-import controllers.front.FrontJsonFapiLive
 import model.ApplicationContext
 import play.api.libs.ws.WSClient
 import play.api.mvc.ControllerComponents
+import services.fronts.FrontJsonFapiLive
 
 trait FaciaControllers {
   def frontJsonFapiLive: FrontJsonFapiLive
